# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Franchise Generation API
  description: |
    REST API for the Franchise Generation platform — a multi-brand franchise ecosystem
    connecting brands, candidates, investors, consultants, and real estate developers.

    ## Authentication
    All protected endpoints require a JWT Bearer token obtained from `POST /auth/login`.
    Include the token in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```

    ## Base URL
    Production: `https://api.franchisegeneration.be/api/v1`
    Local dev: `http://localhost:3000/api/v1`

    ## Roles
    | Role | Description |
    |------|-------------|
    | `admin` | Full backoffice access |
    | `investor` | Investment portal access |
    | `candidate` | Candidate onboarding portal |
    | `consultant` | Lead management access |
    | `developer` | Real estate developer access |
    | `brand` | Brand management portal |

  version: 1.0.0
  contact:
    name: Franchise Generation
    url: https://franchisegeneration.be
servers:
  - url: /api/v1
    description: Current version
security:
  - BearerAuth: []
...
paths:
...
  /candidates/{id}/leads:
    get:
      tags: [Candidates]
      summary: Get candidate's leads
      description: Returns all opportunity leads linked to a candidate.
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Array of candidate leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CandidateLead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Candidates]
      summary: Add lead to candidate
      description: Links a candidate to a new opportunity (creates a CandidateLead record).
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [opportunityId]
              properties:
                opportunityId:
                  type: string
                  example: op-bxl-cookies
                consultantId:
                  type: string
                  description: Override auto-assignment
      responses:
        '201':
          description: Lead created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateLead'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /candidates/{id}/leads/{leadId}:
    put:
      tags: [Candidates]
      summary: Update a candidate lead
      description: Update validation status, fit score, or consultant notes on a specific candidate-opportunity link.
      parameters:
        - $ref: '#/components/parameters/id'
        - name: leadId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                validationStatus:
                  type: string
                  enum: [to-analyse, in-discussion, recommended, validated, rejected, on-hold]
                fitScore:
                  type: integer
                  minimum: 0
                  maximum: 100
                consultantNotes:
                  type: string
      responses:
        '200':
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateLead'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /leads:
    get:
      tags: [Leads]
      summary: List all leads
      description: Returns the full pipeline of candidate leads. Primarily used by the consultant portal.
      parameters:
        - name: brand
          in: query
          schema:
            type: string
        - name: region
          in: query
          schema:
            type: string
        - name: consultant
          in: query
          schema:
            type: string
        - name: status
          in: query
          description: Current pipeline step
          schema:
            type: string
            enum: [interested, consultant-review, first-contact-planned, first-contact-done,
                   financing-precheck, location-validation, business-plan, committee,
                   contract-prep, training-planning, opening-planning]
        - name: type
          in: query
          description: Lead type
          schema:
            type: string
            enum: [franchise, new-brand]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated list of leads
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedList'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Lead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /leads/{id}:
    get:
      tags: [Leads]
      summary: Get lead by ID
      description: Returns the full lead record including history, appointments, and messages.
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Lead detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Leads]
      summary: Update lead
      description: Move a lead to a new pipeline step, update notes, or change priority.
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentStep:
                  type: string
                  enum: [interested, consultant-review, first-contact-planned, first-contact-done,
                         financing-precheck, location-validation, business-plan, committee,
                         contract-prep, training-planning, opening-planning]
                priority:
                  type: string
                  enum: [low, normal, high, urgent]
                notes:
                  type: string
                nextAction:
                  type: string
      responses:
        '200':
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Leads]
      summary: Delete lead
      description: Permanently removes a lead record. Requires admin role.
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Lead deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # =========================================================================
  # CONSULTANTS
  # =========================================================================
...
  /consultants/{id}/leads:
    get:
      tags: [Consultants]
      summary: Get consultant's leads
      description: Returns all leads assigned to a specific consultant.
      parameters:
        - $ref: '#/components/parameters/id'
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Array of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # =========================================================================
  # DEVELOPERS
  # =========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource identifier
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: VALIDATION_ERROR
            message: Request validation failed.
            details:
              - field: email
                message: Must be a valid email address.
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: A valid Bearer token is required.
    Forbidden:
      description: Authenticated but insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: FORBIDDEN
            message: You do not have permission to access this resource.
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NOT_FOUND
            message: The requested resource was not found.
...
  schemas:
...
    CandidateLead:
      type: object
      description: A link between a candidate and a specific opportunity, tracked through the pipeline
      properties:
        id:
          type: string
          example: opp-link-001
        candidateId:
          type: string
          example: cand-001
        opportunityId:
          type: string
          example: op-bxl-cookies
        brand:
          type: string
          example: cookies
        projectType:
          type: string
          example: Kiosque
        location:
          type: string
          example: "Bruxelles · Galerie Toison d'Or"
        validationStatus:
          type: string
          enum: [to-analyse, in-discussion, recommended, validated, rejected, on-hold]
          example: recommended
        fitScore:
          type: integer
          minimum: 0
          maximum: 100
          example: 92
        consultantId:
          type: string
          nullable: true
        consultantNotes:
          type: string
        documents:
          type: array
          items:
            type: string
          description: List of requested/submitted document names
        history:
          type: array
          items:
            type: object
            properties:
              at:
                type: string
                example: "12 mai 2026"
              author:
                type: string
              body:
                type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # -------------------------------------------------------------------------
...
    Lead:
      type: object
      description: A candidate lead in the consultant pipeline
      properties:
        id:
          type: string
          example: lead-001
        candidate:
          type: object
          properties:
            name:
              type: string
            email:
              type: string
            phone:
              type: string
            region:
              type: string
            capital:
              type: string
        opportunityId:
          type: string
        currentStep:
          type: string
          enum: [interested, consultant-review, first-contact-planned, first-contact-done,
                 financing-precheck, location-validation, business-plan, committee,
                 contract-prep, training-planning, opening-planning]
        assignedTo:
          type: object
          properties:
            name:
              type: string
            role:
              type: string
        priority:
          type: string
          enum: [low, normal, high, urgent]
        source:
          type: string
        createdAt:
          type: string
          format: date-time
        lastUpdate:
          type: string
        notes:
          type: string
        nextAction:
          type: string
        requestedDocuments:
          type: array
          items:
            type: string
        appointments:
          type: array
          items:
            type: object
            properties:
              kind:
                type: string
              when:
                type: string
              status:
                type: string
                enum: [planned, done, cancelled]
        messages:
          type: array
          items:
            type: object
            properties:
              from:
                type: string
                enum: [consultant, candidate, admin]
              name:
                type: string
              body:
                type: string
              time:
                type: string
        history:
          type: array
          items:
            type: object
            properties:
              step:
                type: string
              at:
                type: string
              by:
                type: string

    # -------------------------------------------------------------------------
...
    ErrorResponse:
      type: object
      description: Standard error response envelope
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error description
          example: Request validation failed.
        details:
          type: array
          nullable: true
          description: Field-level validation errors
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # -------------------------------------------------------------------------
    PaginatedList:
      type: object
      description: Generic paginated list wrapper
      properties:
        data:
          type: array
          items: {}
          description: Array of result items
        total:
          type: integer
          description: Total number of matching records
          example: 42
        page:
          type: integer
          description: Current page number (1-indexed)
          example: 1
        limit:
          type: integer
          description: Number of items per page
          example: 20
        totalPages:
          type: integer
          description: Total number of pages
          example: 3